# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Primeagent
  version: 1.7.0
paths:
...
  /api/v1/variables/:
    get:
      tags:
      - Variables
      summary: Read Variables
      description: Read all variables.
      operationId: read_variables_api_v1_variables__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/VariableRead'
                type: array
                title: Response Read Variables Api V1 Variables  Get
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
    post:
      tags:
      - Variables
      summary: Create Variable
      description: Create a new variable.
      operationId: create_variable_api_v1_variables__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VariableCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariableRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
  /api/v1/variables/{variable_id}:
    patch:
      tags:
      - Variables
      summary: Update Variable
      description: Update a variable.
      operationId: update_variable_api_v1_variables__variable_id__patch
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: variable_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Variable Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VariableUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VariableRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Variables
      summary: Delete Variable
      description: Delete a variable.
      operationId: delete_variable_api_v1_variables__variable_id__delete
      security:
      - OAuth2PasswordBearer: []
      - API key query: []
      - API key header: []
      parameters:
      - name: variable_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Variable Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VariableCreate:
      properties:
        name:
          type: string
          title: Name
          description: Name of the variable
        value:
          type: string
          title: Value
          description: Encrypted value of the variable
        default_fields:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Default Fields
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
          description: Type of the variable
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
          description: Creation time of the variable
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
          description: Creation time of the variable
      type: object
      required:
      - name
      - value
      - default_fields
      title: VariableCreate
    VariableRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          description: Name of the variable
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
          description: Type of the variable
        value:
          anyOf:
          - type: string
          - type: 'null'
          title: Value
          description: Encrypted value of the variable
        default_fields:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Default Fields
          description: Default fields for the variable
      type: object
      required:
      - id
      title: VariableRead
    VariableUpdate:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          description: Name of the variable
        value:
          anyOf:
          - type: string
          - type: 'null'
          title: Value
          description: Encrypted value of the variable
        default_fields:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Default Fields
          description: Default fields for the variable
      type: object
      required:
      - id
      title: VariableUpdate
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: api/v1/login
    API key query:
      type: apiKey
      in: query
      name: x-api-key
    API key header:
      type: apiKey
      in: header
      name: x-api-key